Exception handling

+++ src/main/java/com/project/ecommerce/services/ProductService.java
@@ -3,16 +3,17 @@ package com.project.ecommerce.services;
 import com.project.ecommerce.dto.ProductDTO;
 import com.project.ecommerce.entities.Product;
 import com.project.ecommerce.repositories.ProductRepository;
+import com.project.ecommerce.services.exceptions.DatabaseException;
 import com.project.ecommerce.services.exceptions.ResourceNotFoundException;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.RequestBody;
-
-import java.util.List;
 
 @Service
 public class ProductService {
@@ -41,8 +42,7 @@ public class ProductService {
     }
 
     @Transactional
-    public ProductDTO update(Long id, ProductDTO dto) {
-
+    public ProductDTO update(Long id, ProductDTO dto) throws ResourceNotFoundException {
         try {
             Product entity = repository.getReferenceById(id);
             copyDtoToEntity(dto,entity);
@@ -54,9 +54,16 @@ public class ProductService {
 
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
